refactor: move ControlledViewer docs files position
